requirements.txt
datasketch
faiss_gpu==1.7.2
func_timeout==4.3.5
gensim
loguru==0.7.2
nltk==3.9.1
numpy
pandas==2.2.3
pyserini==0.44.0
rapidfuzz==3.12.2
Requests==2.32.3
sentence_transformers==2.7.0
spacy==3.7.5
SQLAlchemy==2.0.36
sqlglot==20.11.0
torch==2.5.1
tqdm==4.66.5
transformers==4.48.3
pydantic_settings